Read species labels from a file

This is a convenience function to read species labels from a file.

Usage

get_species_from_file(species_file)

Arguments

species_file

Path to species file.

Value

A vector with class labels e.g. c("Cyanocitta cristata_Blue Jay", "Zenaida macroura_Mourning Dove")

See also

Examples

# Read a custom species file
get_species_from_file(system.file("extdata", "species_list.txt", package = "birdnetR"))
#>  [1] "Accipiter cooperii_Cooper's Hawk"             
#>  [2] "Agelaius phoeniceus_Red-winged Blackbird"     
#>  [3] "Anas platyrhynchos_Mallard"                   
#>  [4] "Anas rubripes_American Black Duck"            
#>  [5] "Ardea herodias_Great Blue Heron"              
#>  [6] "Baeolophus bicolor_Tufted Titmouse"           
#>  [7] "Branta canadensis_Canada Goose"               
#>  [8] "Bucephala albeola_Bufflehead"                 
#>  [9] "Bucephala clangula_Common Goldeneye"          
#> [10] "Buteo jamaicensis_Red-tailed Hawk"            
#> [11] "Cardinalis cardinalis_Northern Cardinal"      
#> [12] "Certhia americana_Brown Creeper"              
#> [13] "Colaptes auratus_Northern Flicker"            
#> [14] "Columba livia_Rock Pigeon"                    
#> [15] "Corvus brachyrhynchos_American Crow"          
#> [16] "Corvus corax_Common Raven"                    
#> [17] "Cyanocitta cristata_Blue Jay"                 
#> [18] "Cygnus olor_Mute Swan"                        
#> [19] "Dryobates pubescens_Downy Woodpecker"         
#> [20] "Dryobates villosus_Hairy Woodpecker"          
#> [21] "Dryocopus pileatus_Pileated Woodpecker"       
#> [22] "Eremophila alpestris_Horned Lark"             
#> [23] "Haemorhous mexicanus_House Finch"             
#> [24] "Haemorhous purpureus_Purple Finch"            
#> [25] "Haliaeetus leucocephalus_Bald Eagle"          
#> [26] "Junco hyemalis_Dark-eyed Junco"               
#> [27] "Larus argentatus_Herring Gull"                
#> [28] "Larus delawarensis_Ring-billed Gull"          
#> [29] "Lophodytes cucullatus_Hooded Merganser"       
#> [30] "Melanerpes carolinus_Red-bellied Woodpecker"  
#> [31] "Meleagris gallopavo_Wild Turkey"              
#> [32] "Melospiza melodia_Song Sparrow"               
#> [33] "Mergus merganser_Common Merganser"            
#> [34] "Mergus serrator_Red-breasted Merganser"       
#> [35] "Passer domesticus_House Sparrow"              
#> [36] "Poecile atricapillus_Black-capped Chickadee"  
#> [37] "Regulus satrapa_Golden-crowned Kinglet"       
#> [38] "Sialia sialis_Eastern Bluebird"               
#> [39] "Sitta canadensis_Red-breasted Nuthatch"       
#> [40] "Sitta carolinensis_White-breasted Nuthatch"   
#> [41] "Spinus pinus_Pine Siskin"                     
#> [42] "Spinus tristis_American Goldfinch"            
#> [43] "Spizelloides arborea_American Tree Sparrow"   
#> [44] "Sturnus vulgaris_European Starling"           
#> [45] "Thryothorus ludovicianus_Carolina Wren"       
#> [46] "Turdus migratorius_American Robin"            
#> [47] "Zenaida macroura_Mourning Dove"               
#> [48] "Zonotrichia albicollis_White-throated Sparrow"

# To access all class labels that are supported in your language,
# you can read in the respective label file
labels_path <- get_labels_path("fr")
species_list <- get_species_from_file(labels_path)
head(species_list)
#> [1] "Abroscopus albogularis_Bouscarle à moustaches"       
#> [2] "Abroscopus schisticeps_Bouscarle à face noire"       
#> [3] "Abroscopus superciliaris_Bouscarle à sourcils blancs"
#> [4] "Aburria aburri_Pénélope aburri"                      
#> [5] "Acanthagenys rufogularis_Méliphage à bavette"        
#> [6] "Acanthidops bairdi_Bec-en-cheville gris"