Check if an Object is a Valid Species List

This internal function checks if an object is either a character vector of length greater than 0 or a list where each element is a single non-empty character string.

Usage

is_valid_species_list(obj)

Arguments

obj

The object to check. This can be either a character vector or a list.

Value

A logical value indicating whether the object is a valid species list:

TRUE if the object is a character vector of length > 0 or a list with each element being a single character string.

FALSE otherwise.

Note

This function is intended for internal use and may not be exported or accessible for external users.

Examples

if (FALSE) { # \dontrun{
is_valid_species_list(c("species1", "species2")) # TRUE
is_valid_species_list(list("species1", "species2")) # TRUE
is_valid_species_list(c(1, 2, 3)) # FALSE
is_valid_species_list(list(a = 1, b = 2)) # FALSE
} # }