Dynamically create a BirdNET model

This function dynamically creates a BirdNET model based on the provided model name and version. It retrieves the appropriate Python model constructor from the module map, evaluates the constructor, and returns a wrapped BirdNET model object.

Usage

model_factory(model_name, version, ...)

Arguments

model_name

Character. The name of the model to create (e.g., "tflite", "protobuf").

version

Character. The version of the model (e.g., "v2.4").

...

Additional arguments passed to the Python model constructor (e.g., tflite_num_threads, language).

Value

A BirdNET model object of class birdnet_model and its subclasses (e.g., "tflite_v2.4").

Examples

if (FALSE) { # interactive()
py_birdnet_models <- reticulate::import("birdnet.models")
birdnet_model <- model_factory("tflite", "v2.4", tflite_num_threads = 2, language = "en_us")
}