Create a new BirdNET model object

This function creates a new BirdNET model object by wrapping a Python model object and assigning it a class and optional subclass. The model is created as an R object that can be interacted with using R's S3 method dispatch.

Usage

new_birdnet_model(x, ..., subclass = character())

Arguments

x

A Python object representing the BirdNET model. This is typically a Python model object created using the reticulate package.

...

Additional attributes to attach to the BirdNET model object.

subclass

Character. An optional subclass name for the BirdNET model (e.g., "tflite_v2.4"). The subclass is combined with the base class birdnet_model.

Value

An S3 object of class birdnet_model (and any specified subclass) containing the Python model object and any additional attributes passed in ....

Examples

if (FALSE) { # interactive()
py_birdnet_models <- reticulate::import("birdnet.models")
tflite_model <- py_birdnet_models$v2m4$AudioModelV2M4TFLite()
birdnet_model <- new_birdnet_model(tflite_model, language = "en_us", version = "v2.4")
}