Predict Species Within an Audio File

This function predicts species within an audio file using the BirdNET model.

Usage

predict_species(
  model,
  audio_file = system.file("extdata", "soundscape.wav", package = "birdnetR"),
  min_confidence = 0.1,
  batch_size = 1L,
  use_bandpass = TRUE,
  bandpass_fmin = 0L,
  bandpass_fmax = 15000L,
  apply_sigmoid = TRUE,
  sigmoid_sensitivity = 1,
  filter_species = NULL,
  file_splitting_duration_s = 600,
  keep_empty = TRUE
)

Arguments

model

BirdNETModel. An instance of the BirdNET model returned by init_model().

audio_file

character. The path to the audio file.

min_confidence

numeric. Minimum confidence threshold for predictions.

batch_size

integer. Number of audio samples to process in a batch.

use_bandpass

logical. Whether to apply a bandpass filter.

bandpass_fmin, bandpass_fmax

numeric. Minimum/Maximum frequency for the bandpass filter (in Hz). Ignored if use_bandpass is False.

apply_sigmoid

logical. Whether to apply a sigmoid function to the model output.

sigmoid_sensitivity

numeric. Sensitivity parameter for the sigmoid function. Must be in the interval 0.5 - 1.5. Ignored if apply_sigmoid is False.

filter_species

NULL, a character vector of length greater than 0 or a list where each element is a single non-empty character string. Used to filter the predictions. If NULL, no filtering is applied. See get_species_from_file() for more details.

file_splitting_duration_s

numeric. Duration in seconds for splitting the audio file into smaller segments for processing.

keep_empty

logical. Whether to include empty intervals in the output.

Value

A data frame with columns: start, end, scientific_name, common_name, and confidence. Each row represents a single prediction.

Details

Applying a sigmoid activation function, (apply_sigmoid=True) scales the unbound class output of the linear classifier ("logit score") to the range 0-1. This confidence score is a unitless, numeric expression of BirdNET’s “confidence” in its prediction (but not the probability of species presence). Sigmoid sensitivity < 1 leads to more higher and lower scoring predictions and a value > 1 leads to more intermediate-scoring predictions.

For more information on BirdNET confidence scores, the sigmoid activation function and a suggested workflow on how to convert confidence scores to probabilities, see Wood & Kahl, 2024

References

Wood, C. M., & Kahl, S. (2024). Guidelines for appropriate use of BirdNET scores and other detector outputs. Journal of Ornithology. https://doi.org/10.1007/s10336-024-02144-5

See also