Predict species for a given location and time

Uses the BirdNET Species Range Model to estimate the presence of bird species at a specified location and time of year.

Usage

predict_species_at_location_and_time(
  model,
  latitude,
  longitude,
  week = NULL,
  min_confidence = 0.03
)

Arguments

model

BirdNETModel. An instance of the BirdNET model returned by init_model().

latitude

numeric. The latitude of the location for species prediction. Must be in the interval [-90.0, 90.0].

longitude

numeric. The longitude of the location for species prediction. Must be in the interval [-180.0, 180.0].

week

integer. The week of the year for which to predict species. Must be in the interval [1, 48] if specified. If NULL, predictions are not limited to a specific week.

min_confidence

numeric. Minimum confidence threshold for predictions to be considered valid. Must be in the interval [0, 1.0).

Value

A data frame with columns: label, confidence. Each row represents a predicted species, with the confidence indicating the likelihood of the species being present at the specified location and time.

Details

The BirdNET Species Range Model leverages eBird checklist frequency data to estimate the probability of bird species occurrences based on latitude, longitude, and time of year. It integrates actual observations and expert-curated data, making it adaptable to regions with varying levels of data availability. The model employs circular embeddings and a classifier to predict species presence and migration patterns, achieving higher accuracy in data-rich regions and lower accuracy in underrepresented areas like parts of Africa and Asia. For more details, you can view the full discussion here: https://github.com/kahst/BirdNET-Analyzer/discussions/234

Examples

# Predict species in Chemnitz, Germany, that are present all year round
model <- init_model(language = "de")
predict_species_at_location_and_time(model, latitude = 50.8334, longitude = 12.9231)
#>                                           label confidence
#> 1                         Parus major_Kohlmeise 0.99645239
#> 2                           Turdus merula_Amsel 0.99553460
#> 3                  Anas platyrhynchos_Stockente 0.98629630
#> 4                  Columba palumbus_Ringeltaube 0.97970045
#> 5                 Cyanistes caeruleus_Blaumeise 0.97704250
#> 6                    Fringilla coelebs_Buchfink 0.97502279
#> 7                Erithacus rubecula_Rotkehlchen 0.95230711
#> 8            Sylvia atricapilla_Mönchsgrasmücke 0.94955504
#> 9               Phylloscopus collybita_Zilpzalp 0.94362831
#> 10                    Motacilla alba_Bachstelze 0.93249565
#> 11                Hirundo rustica_Rauchschwalbe 0.92767054
#> 12              Garrulus glandarius_Eichelhäher 0.92274117
#> 13                     Buteo buteo_Mäusebussard 0.91962087
#> 14                 Dendrocopos major_Buntspecht 0.91961437
#> 15                        Apus apus_Mauersegler 0.91489345
#> 16                        Sturnus vulgaris_Star 0.91195983
#> 17                             Pica pica_Elster 0.89453363
#> 18                Carduelis carduelis_Stieglitz 0.88957518
#> 19                     Ardea cinerea_Graureiher 0.88793421
#> 20                Emberiza citrinella_Goldammer 0.88048667
#> 21            Troglodytes troglodytes_Zaunkönig 0.87095416
#> 22                Turdus philomelos_Singdrossel 0.87011319
#> 23                       Sitta europaea_Kleiber 0.87003815
#> 24                     Chloris chloris_Grünfink 0.86355007
#> 25               Passer domesticus_Haussperling 0.86212170
#> 26                     Cygnus olor_Höckerschwan 0.85587966
#> 27                     Corvus corone_Rabenkrähe 0.85445541
#> 28                Delichon urbicum_Mehlschwalbe 0.83508080
#> 29                        Fulica atra_Blässhuhn 0.82934046
#> 30                     Picus viridis_Grünspecht 0.82888216
#> 31                  Falco tinnunculus_Turmfalke 0.82351327
#> 32              Turdus pilaris_Wacholderdrossel 0.82054323
#> 33                 Phalacrocorax carbo_Kormoran 0.81945527
#> 34          Phoenicurus ochruros_Hausrotschwanz 0.80881274
#> 35                 Passer montanus_Feldsperling 0.78537780
#> 36                         Anser anser_Graugans 0.78349972
#> 37          Chroicocephalus ridibundus_Lachmöwe 0.76751101
#> 38                   Aythya fuligula_Reiherente 0.75682569
#> 39             Podiceps cristatus_Haubentaucher 0.75553709
#> 40                   Alauda arvensis_Feldlerche 0.69698489
#> 41                    Spinus spinus_Erlenzeisig 0.69496942
#> 42            Streptopelia decaocto_Türkentaube 0.68757200
#> 43                    Columba livia_Felsentaube 0.68738079
#> 44                      Cuculus canorus_Kuckuck 0.67729229
#> 45                      Ardea alba_Silberreiher 0.65753788
#> 46                 Phylloscopus trochilus_Fitis 0.65112913
#> 47                        Anas crecca_Krickente 0.61713117
#> 48                    Lanius collurio_Neuntöter 0.61520779
#> 49                Mareca strepera_Schnatterente 0.61060512
#> 50                        Corvus corax_Kolkrabe 0.60739279
#> 51          Tachybaptus ruficollis_Zwergtaucher 0.60032487
#> 52                  Mergus merganser_Gänsesäger 0.59896916
#> 53               Emberiza schoeniclus_Rohrammer 0.58444989
#> 54     Coccothraustes coccothraustes_Kernbeißer 0.58386409
#> 55                 Poecile palustris_Sumpfmeise 0.58308756
#> 56      Acrocephalus scirpaceus_Teichrohrsänger 0.58198237
#> 57                      Aythya ferina_Tafelente 0.57819277
#> 58                       Milvus milvus_Rotmilan 0.57267773
#> 59               Curruca communis_Dorngrasmücke 0.56269276
#> 60           Regulus regulus_Wintergoldhähnchen 0.55604601
#> 61                     Pyrrhula pyrrhula_Gimpel 0.54912925
#> 62               Linaria cannabina_Bluthänfling 0.53836346
#> 63                        Corvus monedula_Dohle 0.53744876
#> 64             Luscinia megarhynchos_Nachtigall 0.53165764
#> 65                   Periparus ater_Tannenmeise 0.53161097
#> 66             Aegithalos caudatus_Schwanzmeise 0.53160346
#> 67                Phasianus colchicus_Jagdfasan 0.53132844
#> 68                      Serinus serinus_Girlitz 0.52792656
#> 69                    Vanellus vanellus_Kiebitz 0.52776206
#> 70              Turdus viscivorus_Misteldrossel 0.51759207
#> 71              Dryocopus martius_Schwarzspecht 0.49939877
#> 72                  Motacilla flava_Schafstelze 0.49885574
#> 73       Certhia brachydactyla_Gartenbaumläufer 0.49876487
#> 74                Gallinula chloropus_Teichhuhn 0.49523902
#> 75                       Alcedo atthis_Eisvogel 0.48993650
#> 76                   Ciconia ciconia_Weißstorch 0.48821187
#> 77                 Sylvia borin_Gartengrasmücke 0.47310179
#> 78             Curruca curruca_Klappergrasmücke 0.47287306
#> 79                        Oriolus oriolus_Pirol 0.47245672
#> 80                 Circus aeruginosus_Rohrweihe 0.47114873
#> 81           Actitis hypoleucos_Flussuferläufer 0.46899500
#> 82                  Corvus frugilegus_Saatkrähe 0.46493271
#> 83           Prunella modularis_Heckenbraunelle 0.46352315
#> 84                 Alopochen aegyptiaca_Nilgans 0.45838946
#> 85       Acrocephalus palustris_Sumpfrohrsänger 0.45668471
#> 86     Phoenicurus phoenicurus_Gartenrotschwanz 0.44974494
#> 87                  Spatula clypeata_Löffelente 0.44431674
#> 88                      Accipiter nisus_Sperber 0.44304466
#> 89                Bucephala clangula_Schellente 0.43756390
#> 90                            Grus grus_Kranich 0.43635592
#> 91            Tringa glareola_Bruchwasserläufer 0.43420565
#> 92            Lophophanes cristatus_Haubenmeise 0.42019022
#> 93                    Mareca penelope_Pfeifente 0.41281414
#> 94              Sterna hirundo_Flussseeschwalbe 0.40616271
#> 95                Gallinago gallinago_Bekassine 0.40309733
#> 96                  Milvus migrans_Schwarzmilan 0.39692584
#> 97  Acrocephalus arundinaceus_Drosselrohrsänger 0.39124268
#> 98            Fringilla montifringilla_Bergfink 0.39039764
#> 99              Muscicapa striata_Grauschnäpper 0.37686411
#> 100         Charadrius dubius_Flussregenpfeifer 0.37177074
#> 101             Motacilla cinerea_Gebirgsstelze 0.37119526
#> 102                Spatula querquedula_Knäkente 0.36937600
#> 103                Larus cachinnans_Steppenmöwe 0.36550432
#> 104               Anthus pratensis_Wiesenpieper 0.36459562
#> 105               Tringa nebularia_Grünschenkel 0.35800940
#> 106                Branta canadensis_Kanadagans 0.35513273
#> 107                Riparia riparia_Uferschwalbe 0.34158945
#> 108           Certhia familiaris_Waldbaumläufer 0.34057742
#> 109                 Calidris pugnax_Kampfläufer 0.33378395
#> 110                 Anthus trivialis_Baumpieper 0.32977489
#> 111                Poecile montanus_Weidenmeise 0.32193810
#> 112      Regulus ignicapilla_Sommergoldhähnchen 0.31415448
#> 113                   Anser albifrons_Blässgans 0.31123361
#> 114                     Columba oenas_Hohltaube 0.31031001
#> 115                   Turdus iliacus_Rotdrossel 0.30680388
#> 116                 Emberiza calandra_Grauammer 0.30615717
#> 117            Tringa ochropus_Waldwasserläufer 0.30092895
#> 118              Hippolais icterina_Gelbspötter 0.28836453
#> 119                       Larus canus_Sturmmöwe 0.28737485
#> 120                     Netta rufina_Kolbenente 0.28517705
#> 121                    Corvus cornix_Nebelkrähe 0.28454724
#> 122              Saxicola rubetra_Braunkehlchen 0.28386748
#> 123 Acrocephalus schoenobaenus_Schilfrohrsänger 0.28029752
#> 124      Loxia curvirostra_Fichtenkreuzschnabel 0.27461991
#> 125      Phylloscopus sibilatrix_Waldlaubsänger 0.26411819
#> 126          Ficedula hypoleuca_Trauerschnäpper 0.26059541
#> 127           Saxicola rubicola_Schwarzkehlchen 0.26029003
#> 128                Rallus aquaticus_Wasserralle 0.25950387
#> 129         Locustella luscinioides_Rohrschwirl 0.25636688
#> 130           Calidris alpina_Alpenstrandläufer 0.25567725
#> 131           Anser serrirostris_Tundrasaatgans 0.24914101
#> 132               Locustella naevia_Feldschwirl 0.24222459
#> 133               Haliaeetus albicilla_Seeadler 0.23829913
#> 134                        Anas acuta_Spießente 0.23819828
#> 135                 Dryobates minor_Kleinspecht 0.23068380
#> 136        Tringa erythropus_Dunkelwasserläufer 0.22931644
#> 137            Dendrocoptes medius_Mittelspecht 0.22652707
#> 138                Anthus spinoletta_Bergpieper 0.22452772
#> 139            Larus michahellis_Mittelmeermöwe 0.21748297
#> 140                    Falco subbuteo_Baumfalke 0.21098644
#> 141                  Accipiter gentilis_Habicht 0.20669085
#> 142                 Lullula arborea_Heidelerche 0.20550326
#> 143             Streptopelia turtur_Turteltaube 0.20004660
#> 144                 Cinclus cinclus_Wasseramsel 0.19380097
#> 145                 Numenius arquata_Brachvogel 0.19191389
#> 146                        Strix aluco_Waldkauz 0.18986404
#> 147                    Circus cyaneus_Kornweihe 0.18354732
#> 148            Oenanthe oenanthe_Steinschmätzer 0.18338126
#> 149                Pandion haliaetus_Fischadler 0.18336481
#> 150               Pernis apivorus_Wespenbussard 0.17668539
#> 151          Chlidonias niger_Trauerseeschwalbe 0.17596099
#> 152               Luscinia svecica_Blaukehlchen 0.17522924
#> 153                 Lanius excubitor_Raubwürger 0.16994186
#> 154     Podiceps nigricollis_Schwarzhalstaucher 0.16804053
#> 155                    Jynx torquilla_Wendehals 0.16763055
#> 156                   Tadorna tadorna_Brandgans 0.16476496
#> 157           Podiceps grisegena_Rothalstaucher 0.16474454
#> 158         Nucifraga caryocatactes_Tannenhäher 0.16452831
#> 159                  Tringa totanus_Rotschenkel 0.15846707
#> 160               Mergellus albellus_Zwergsäger 0.15125172
#> 161                      Picus canus_Grauspecht 0.15034559
#> 162                 Ciconia nigra_Schwarzstorch 0.14960133
#> 163       Charadrius hiaticula_Sandregenpfeifer 0.14472079
#> 164           Nycticorax nycticorax_Nachtreiher 0.13774449
#> 165                Falco peregrinus_Wanderfalke 0.13579598
#> 166               Botaurus stellaris_Rohrdommel 0.13425295
#> 167               Aix galericulata_Mandarinente 0.13419616
#> 168                Remiz pendulinus_Beutelmeise 0.12373128
#> 169                 Panurus biarmicus_Bartmeise 0.12137543
#> 170                   Coturnix coturnix_Wachtel 0.11635761
#> 171                       Perdix perdix_Rebhuhn 0.11422057
#> 172                Turdus torquatus_Ringdrossel 0.11380927
#> 173             Branta leucopsis_Weißwangengans 0.11346602
#> 174               Egretta garzetta_Seidenreiher 0.11295213
#> 175                       Upupa epops_Wiedehopf 0.10721239
#> 176                    Larus fuscus_Heringsmöwe 0.10558451
#> 177                  Limosa limosa_Uferschnepfe 0.10455787
#> 178        Pluvialis apricaria_Goldregenpfeifer 0.09768523
#> 179                      Aythya marila_Bergente 0.09661443
#> 180                 Buteo lagopus_Raufußbussard 0.09334701
#> 181           Calidris minuta_Zwergstrandläufer 0.09098604
#> 182                       Asio otus_Waldohreule 0.09062644
#> 183               Merops apiaster_Bienenfresser 0.09046380
#> 184                 Platalea leucorodia_Löffler 0.08989146
#> 185                 Circus pygargus_Wiesenweihe 0.08737607
#> 186                 Larus argentatus_Silbermöwe 0.08355612
#> 187    Pluvialis squatarola_Kiebitzregenpfeifer 0.08350678
#> 188             Scolopax rusticola_Waldschnepfe 0.08286890
#> 189               Podiceps auritus_Ohrentaucher 0.08090969
#> 190              Ixobrychus minutus_Zwergdommel 0.07995533
#> 191             Pyrrhocorax graculus_Alpendohle 0.07927271
#> 192                 Ardea purpurea_Purpurreiher 0.07800952
#> 193       Recurvirostra avosetta_Säbelschnäbler 0.07763951
#> 194      Calidris ferruginea_Sichelstrandläufer 0.07641439
#> 195  Ichthyaetus melanocephalus_Schwarzkopfmöwe 0.07582787
#> 196                 Gavia stellata_Sterntaucher 0.07385838
#> 197                    Melanitta fusca_Samtente 0.06954328
#> 198        Locustella fluviatilis_Schlagschwirl 0.06944773
#> 199           Numenius phaeopus_Regenbrachvogel 0.06751283
#> 200       Ptyonoprogne rupestris_Felsenschwalbe 0.06720366
#> 201              Hydrocoloeus minutus_Zwergmöwe 0.06686041
#> 202       Ficedula albicollis_Halsbandschnäpper 0.06679752
#> 203         Himantopus himantopus_Stelzenläufer 0.06530610
#> 204           Plectrophenax nivalis_Schneeammer 0.06238731
#> 205                Aquila chrysaetos_Steinadler 0.06201217
#> 206                     Athene noctua_Steinkauz 0.06179736
#> 207                 Gavia arctica_Prachttaucher 0.06144491
#> 208              Cygnus columbianus_Zwergschwan 0.06139188
#> 209                     Bubulcus ibis_Kuhreiher 0.06134326
#> 210             Porzana porzana_Tüpfelsumpfhuhn 0.06016217
#> 211              Galerida cristata_Haubenlerche 0.05939114
#> 212     Chlidonias hybrida_Weißbart-Seeschwalbe 0.05856375
#> 213                    Cygnus cygnus_Singschwan 0.05591480
#> 214         Sternula albifrons_Zwergseeschwalbe 0.05400210
#> 215                 Mergus serrator_Mittelsäger 0.05336181
#> 216                      Apus melba_Alpensegler 0.05325992
#> 217                    Falco columbarius_Merlin 0.05301178
#> 218                      Crex crex_Wachtelkönig 0.05109008
#> 219          Lymnocryptes minimus_Zwergschnepfe 0.05011262
#> 220            Curruca nisoria_Sperbergrasmücke 0.04815649
#> 221          Acanthis flammea_Taigabirkenzeisig 0.04760549
#> 222         Caprimulgus europaeus_Nachtschwalbe 0.04744333
#> 223                  Melanitta nigra_Trauerente 0.04599626
#> 224                 Clanga pomarina_Schreiadler 0.04582243
#> 225                      Aythya nyroca_Moorente 0.04459698
#> 226              Limosa lapponica_Pfuhlschnepfe 0.04398722
#> 227        Haematopus ostralegus_Austernfischer 0.04346076
#> 228                      Calidris canutus_Knutt 0.04323935
#> 229                  Emberiza hortulana_Ortolan 0.04299862
#> 230              Arenaria interpres_Steinwälzer 0.04243961
#> 231                  Asio flammeus_Sumpfohreule 0.04041028
#> 232               Anthus campestris_Brachpieper 0.03993680
#> 233             Cisticola juncidis_Zistensänger 0.03991739
#> 234          Acanthis cabaret_Alpenbirkenzeisig 0.03984652
#> 235            Prunella collaris_Alpenbraunelle 0.03751016
#> 236    Thalasseus sandvicensis_Brandseeschwalbe 0.03711388
#> 237         Glaucidium passerinum_Sperlingskauz 0.03687109
#> 238          Psittacula krameri_Halsbandsittich 0.03669582
#> 239           Bombycilla garrulus_Seidenschwanz 0.03285863
#> 240    Calidris temminckii_Temminckstrandläufer 0.03176918
#> 241          Carpodacus erythrinus_Karmingimpel 0.03075498
#> 242               Zapornia parva_Kleinsumpfhuhn 0.03056428